+++ hr-worker/src/main/java/com/gabriel/hrworker/repositories/WorkerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gabriel.hrworker.entities.Worker;
 
-public interface WorkerRepository extends JpaRepository<Worker, Integer> {
+public interface WorkerRepository extends JpaRepository<Worker, Long> {
 
 }
